fix: extract auto pause functional to custom hook in bg-video widget
